Create README.md
What Github is doing with their gitcoin, is absolutely what this community needs.  Rewarding those who not only provide more to the open source code development, but that also add awareness to this growing ecosystem of cryptography and decentralized information / value, via art, podcast, blogs etc. I believe will further expand upon the network effect needed for a digital trending world to move to, and work on blockchain tech as the bedrock for a better society.